Additional fixes for namespaces, and for single elements within a namespace.
Restrict some namespace behaviour that is not permitted (use of xml, xmlns
namespaces). Return an error if a no namespace element is used and the default
namespace is not no-namespace (there is no other way to specify it). Fix a
problem where a single simple tag within a namespace would not set the default
namespace, becuase ClosePendingNamespace (and hence OpenPendingNamespacePrefix)
was called after the namespace had reverted to its previous value.
been reverted to the old namespace.