Make workflows compatible with Workbench by default
The t2flow format allows activity input and output ports to be directly mapped
to alternative names. This is of most use for RShell, where an original R
script contains a name including a dot (e.g. x.y is a valid identifier in R,
similar to x_y in most other languages).

Balcazapy supports this mapping, and it runs in Taverna Engine. However,
Workbench does not show links to dotted names, and cannot edit the RShell
while the dotted names exist. It wants to use the activity name for the
Taverna process name, and the dotted name is not a valid Taverna process name.

This commit changes the behaviour to add lines to the start and end of an
RShell activity to do the mapping within the R script. So, the activity and
process names are identical (and compatible with Workbench). And the R script
is modified to map to the correct name.

The R script has new names added, but these are the names of Taverna process
ports, so cannot clash with other ports. If there is a variable with the
same name as the Taverna port, it should not cause problems, as the variable
must be set after the mapping (for inputs), or completed before the mapping
(for outputs). (Except for some perverse cross-mappings between names).

The only problem area might be if the Taverna name clashes with a built-in
R name (e.g. user might conceivablly map task.input.list[l.is.t], which would
attempt to add the line "l.is.t <- list" to the start of the script.