Change activity assignment to use left-shift
Using = to map activities to tasks can be confusing, since it is similar to
setting variables, which is useufl for tasks and activities.

For example, to reuse an activity, you might say
MyActivity = BeanShell(...)

You can then map a task using:
flow.task.MyTask = MyActivity

To make the task input and output mapping simpler, you can assign the task
to a variable.

MyTask = flow.task.MyTask
... >> MyTask.input.in1

Given two assigments like this, it is Pythonic to say:
MyTask = flow.task.MyTask = MyActivity

But this does not do the same thing as the two separate statements. MyTask
ends up being assigned the activity, not the task.

Introducing << for assignment of activities to tasks allows the right thing
to happen:
MyTask = flow.task.MyTask = MyActivity
... >> MyTask.input.in1

The original assignment still exists, and the implementation of << just
calls setattr (the implementation for =). Eventually, we can restructure
properly, once all the extant code has been modified.