Minor quote changes for workspace identifier
This was done to support a leading underscore for the name, making it less
likely to be used by an existing R script. However, Taverna uses input and
output port names without quotes, so this does not work anyway.  Keeping it
since it took significant time to figure out R's rules for using underscores
in names.

Basically, underscore is often interpreted as <-, for backwards compatibility
with S, so need to quote any name with _ for safety. In expressions (i.e. RHS
of assignment), need to
use get() function, e.g.:
"my_var" <- get("my_var")