Output file vars are set at start of R script
Although files appear as output ports, the variable is set to a valid
pathname at the start of the script, to allow the script to use the
variable to write to a co-ordinated location. Hence, the mapping of names
for output files needs to happen at the start.  It also needs to happen at
the end, to ensure the value is correct on exit.