Add workflow signature command option
Calling "balc --signature ..." provides a description of the workflow as
a nested workflow, with inputs and outputs specified. This is useful if
we add a requirement for nested workflows to specify their inputs and
outputs.  This is a planned modification, as it helps to verify that workflows
in different files remain compatible, and guide the user in the outer
workflow as to what inputs and outputs are available.