Remove colon from producedBy attribute
myExperiment will not allow upload of a t2flow workflow with a colon in the
workflow producedBy attribute. It appears to break its RDF parsing.  So we
shorten the URL in there to use a string with no schema, but which will get
to the correct place if pasted into a web browser, since web browsers
typically assume http if no schema is provided.