Re-load input ports after loading R workspace
If the workspace contained a variable with the same name as an input port,
then the workspace variable won.  This fix changes it to the other way -
an RShell input port should be set rather than the workspace variable.

It does this by saving the current R workspace containing the input
ports, loading the input R workspace, and then loading the saved
input ports over the top.